Move compilation steps into own files
