+++ crates/compiler/src/compilation_steps.rs
@@ -1,0 +1,16 @@
+mod add_initial_value_registrations;
+mod add_tracking_declarations;
+mod check_types;
+mod find_tracking_nodes;
+mod generate_code;
+mod get_declarations;
+mod parse_files;
+mod register_initial_variables;
+mod register_strings;
+mod validate_unique_node_names;
+
+pub(crate) use self::{
+    add_initial_value_registrations::*, add_tracking_declarations::*, check_types::*,
+    find_tracking_nodes::*, generate_code::*, get_declarations::*, parse_files::*,
+    register_initial_variables::*, register_strings::*, validate_unique_node_names::*,
+};
