+++ crates/compiler/src/lib.rs
@@ -1,3 +1,4 @@
+pub(crate) mod compilation_steps;
 pub(crate) mod compiler;
 pub(crate) mod error_strategy;
 mod file_parse_result;
